% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_terminal_edit.R
\name{edit_data_frame_in_terminal}
\alias{edit_data_frame_in_terminal}
\title{Edit data frame in terminal}
\usage{
edit_data_frame_in_terminal(df_to_edit = data.frame(matrix(vector(), 0, 2,
  dimnames = list(c(), c("parameter", "value"))), stringsAsFactors = F),
  action = 0)
}
\arguments{
\item{df_to_edit}{= data frame to edit}

\item{action}{with the following options:
= 1 (add new line)
= 2 (edit existing record)
= 3 (remove existing record)
= 4 (save and exit)}
}
\description{
Edit data frame in terminal
}
\examples{
edit_data_frame_in_terminal(data_frame_to_edit)
}

