% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_terminal_edit.R
\name{edit_one_record_in_terminal}
\alias{edit_one_record_in_terminal}
\title{Edit one record from a data frame in terminal}
\usage{
edit_one_record_in_terminal(record_to_edit = data.frame(parameter = c(""),
  value = c(""), stringsAsFactors = F))
}
\arguments{
\item{record_to_edit}{= record from a data frame to edit}
}
\description{
Edit one record from a data frame in terminal
}

