% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_terminal_edit.R
\name{get_value_from_prompt}
\alias{get_value_from_prompt}
\title{Get value from the terminal prompt}
\usage{
get_value_from_prompt(variable_name = "variable", current_value = "",
  example_value = "Value123", show_result = F)
}
\arguments{
\item{variable_name}{= name of the variable}

\item{current_value}{= the current value}

\item{example_value}{= for example (optional)}
}
\description{
Get value from the terminal prompt
}

